Suggestion: Try applying a change to the file (WIP)
Not working yet, but the idea is to get the file referenced by the
suggestion and apply the suggested change directly to it.

Here, we read the original file, and write its contents to the new file,
replacing the lines modified by the suggestion with the suggested
change.

Currently doesn't work because you can't pass file instances to
`fs::rename`, only paths.

I'd like to try a slightly different system so I can keep the original
file's creation date. With the current system, it would be overwritten
by the new temporary file created to store the change.