Suggestion.apply_to(): Change `from` to be `BufRead`
Take a `BufRead`er instead of a file path to read from. This gives us
more flexibility.

We can't use the `Read` trait because of the error described in
b4a7561ccc1ed7dab6cd2c786902dec8c4ad9be6.

Fix the bug in `Suggestion.diff()` so that the suggestion is applied to
the commit it corresponds to instead of HEAD.