Test `Suggestion.diff()`
Add a test for diff generation. Split `diff()` into `diff_with_repo()`
to allow us to test with a new repo instead of defaulting to the current
one.

In the test, create a new Git repo using `git2`, add and commit a file,
and check that the diff generated by a Suggestion is correct.

Got the `IndexEntry` format from
https://docs.rs/git2/0.13.8/src/git2/index.rs.html#858-871.