Suggestion.apply(): Error when patch doesn't apply
Don't apply the patch when it doesn't apply, instead fail with an error.

Previously, the `apply()` method would apply the patch to the lines from
`Suggestion`, even if the file was different from what the suggestion
expected. So, you could end up with the `Suggestion` lines being
overwritten with the suggestion text, even when they didn't match the
original side of the suggestion.

Fix this by getting the diff text from `diff_with_repo()` and applying
it using `git-apply`. This gives us a suitable error when the patch
doesn't apply to the file, leaving the file alone.