Get config
We need the following values in order to build a `Client`:

* GitHub token
* Repo owner
* Repo name

Get the token from a command-line argument, or else the Git config, or
else an environment variable.

Get the repo identifiers from the repo's remote URL. Use the remote
specified as a command-line argument, otherwise get it from the Git
config, or else default to "origin".

TODO: Only try to get the owner-name pair from the remote if a comment
ID was given, not if a URL argument was passed.