OwnerRepo: Parse from SSH-style paths
Remotes can be HTTP URLs or use SSH/SCP-style paths like:

    git@github.com/teddywing/github-suggestion.git

We need to handle both regular URLs and the above style, which isn't
parseable by the 'url' crate.

Add a custom parsing function to get the owner and repo from this
format.

When trying to parse `OwnerRepo` from a string, first try to parse it as
a URL. If that fails with a `RelativeUrlWithoutBase` error, assume it's
the other style of reference.