git-sugpatch: Accept multiple suggestion args; Accept suggestion ID args
Add the ability to pass multiple suggestions to the binary. The diffs
will appear in the output one after another.

Also allow suggestions to be specified by their comment ID instead of
their full URL. This uses the Git repo's remote to determine the
owner-repo pair. When a URL argument is given, the owner-repo pair is
extracted from the URL.

It's now possible to call the binary like this:

    $ git-sugpatch 459692838 \
    > https://github.com/teddywing/suggestion-test/pull/1#discussion_r459691747