Suggestion.diff_command: Return the right side version blob
Don't run the `git diff` command in this function, instead return the
new blob created for the right hand side of the comparison.

This will facilitate showing the diffs of multiple suggestions at once.
Maybe. I haven't tried the command yet.

Actually, looking at the man page, it doesn't say we can include more
than one blob pair, so I might need to put all of the suggestion changes
into a single blob and compare the original commit with that.