Suggestion.diff_command: Diff an index instead of a file
This will enable us to use the `git diff` command with more than one
suggestion at the same time.

We compare the commit of the suggestion with the new index's tree.

And as I write this, I'm realising that we can't diff multiple
suggestions using the same commit because each suggestion has its own
commit.

Need to work out a different way to do this.