git-sugdiff: Add `--no-pager` to `git diff` command
The diff command wasn't exiting after outputting the diff. Add the
`--no-pager` argument to ensure that the pager doesn't activate. This
also seems to cause the command to exit once all the output is printed.

We don't want the pager because we need to call `git diff` for each
suggestion. If an early diff took up more than a page and activated the
pager, that would block the next diffs from appearing. We want them all
to appear at once.