Rename `git-sugpatch` to `git-sugdiff`
The word "patch" reminds me more of the `patch` Unix command, which is
what `git-sugapply` does.

Since this command outputs diffs, it makes more sense to call it
`sugdiff`. That also brings these closer to the native Git commands,
`git-diff` and `git-apply`.

I had chosen "patch" originally because the command generated a unified
diff that could be used as a patch file to apply to the repo.