git-sugdiff: Fix output by waiting for child process
I thought I'd resolved this problem by adding the `--no-pager` flag in
76079daf1e2175d02c9c584e47a21650fce30bb6, but it turns out that didn't
fix it.

When running the command, I'd get the following output, with the shell
prompt preceding the start of the diff, and the command hanging without
exiting to a new prompt line until a new interaction:

    $ ./target/debug/git-sugdiff 459692838
    $ diff --git a/src/server.rs b/44ff3b520321467a9220965d13f997690ff3fda7
    index d9eee95..44ff3b5 100644
    --- a/src/server.rs
    +++ b/44ff3b520321467a9220965d13f997690ff3fda7
    @@ -73,7 +73,7 @@ impl Server {
                            internal_server_error(&mut raw_request.stdout())
                        },
                        Error::ConduitResponse { .. } => {
                            error!("Error [-getting-]{+from+} response: {}", e);

                            internal_server_error(&mut raw_request.stdout())
                        },

Turns out I need to wait for the command to exit in order to properly
execute the external command.