diff_options: Merge `ARG_OPTIONS` and `OPT_OPTIONS`
Now that I've confirmed that Git's key value options require an '=' to
specify the value, we can drop the distinction between options that
require and argument and those that optionally take an argument.

Also remove the code that interprets an argument that doesn't start with
'-' as the value of the preceding option.